Add Docker container configuration
Created a basic Dockerfile and docker-compose configuration. The user
should only have to deal with mounting the logs and maps directories
into the container and specifying the token, channel and rcon password
using environment variables.

I created a separate configuration file that reads zomboi config values
from environment variables that is copied into the container. This
probably could be implemented more directly so a separate file is not
needed.